@@ -5,19 +5,19 @@ import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 
-@EnableWebSecurity
-@Configuration
-public class WebSecurityConfiguration extends WebSecurityConfigurerAdapter {
 
-    @Override
+public class WebSecurityConfiguration{
+
     protected void configure(HttpSecurity http) throws Exception{
         http.authorizeRequests()
-                .antMatchers("/")
+                .antMatchers("/swagger.ui**")
                 .permitAll()
                 .and()
                 .authorizeRequests()
                 .antMatchers("/h2-console/**")
-                .permitAll();
+                .permitAll()
+        .anyRequest().permitAll()
+                ;
 
         http.csrf().disable();
         http.headers().frameOptions().disable();
